move private api to oonimeasurements
move shared methods to common e.g. generate_report_id
consolidate and move shared test/fixtures/initdb to common/fixtures and add links
