+++ ooniapi/common/fixtures/initdb/04-fastpath.sql
