+++ ooniapi/services/oonimeasurements/pyproject.toml
@@ -17,8 +17,10 @@ dependencies = [
   "sqlalchemy ~= 2.0.27",
   "ujson ~= 5.9.0",
   "urllib3 ~= 2.1.0",
+  "pycountry ~= 26.2.16",
   "python-dateutil ~= 2.8.2",
   "pydantic-settings ~= 2.1.0",
+  "pydantic_extra_types ~= 2.11.1",
   "statsd ~= 4.0.1",
   "uvicorn ~= 0.25.0",
   "httpx ~= 0.26.0",
