+++ ooniapi/services/oonimeasurements/src/oonimeasurements/routers/private.py
@@ -21,15 +21,13 @@ from pydantic_extra_types.country import CountryAlpha2
 from pydantic_extra_types.domain import DomainStr
 from pydantic import AnyUrl, Field, StringConstraints, IPvAnyAddress, BeforeValidator
 
-from .v1.probe_services import probe_geoip, generate_test_helpers_conf
 from ..common.clickhouse_utils import query_click, query_click_one_row
 from ..common.dependencies import role_required, ClickhouseDep
 from ..common.prio import generate_test_list
 from ..common.routers import BaseModel
-from ..metrics import Metrics
-from ..countries import lookup_country
+from ..common.countries import lookup_country
+from ..common.utils import generate_report_id
 from ..data import dnscheck_inputs, stunreachability_inputs
-from ..utils import generate_report_id
 
 
 # The private API is exposed under the prefix /api/_
