+++ ooniapi/services/oonimeasurements/tests/fixtures/initdb
@@ -1,0 +1,1 @@
+../../../../common/fixtures/initdb
