--- ooniapi/services/ooniprobe/tests/fixtures/initdb/init.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-set -e
-
-# Add initialization code here. Example: fetch data, generate it dynamically
